Add and Remove functionality done.

+++ Products/Products/Product.cs
@@ -6,7 +6,7 @@
         {
         }
 
-        public Product(string name,int id, string description, double weight, Size size, bool instock = false )
+        public Product(string name,int id, string description, float weight, Size size, bool instock = false )
         {
             ProductName = name;
             ProductID = id;
@@ -15,11 +15,11 @@
             InStock = instock;
             ProductSize = size;
         }
-
+       
         public string ProductName { get; set; }
         public int ProductID { get; set; }
         public string ProductDescription { get; set; }
-        public double Weight { get; set; }
+        public float Weight { get; set; }
         public bool InStock { get; set; }
 
         public Size ProductSize { get; set; }
